refactor(nn): split model config into a module directory
NetworkConfig+builder and LayerConfig+errors move to their own files
(network.rs / layer.rs) under config/, mod.rs keeping the doc and
re-exports. Pure split, no behavior change — makes room for the
upcoming ModelConfig bundle.
