+++ core/src/nn/model/config/layer.rs
@@ -1,6 +1,4 @@
-//! Layer and network configuration: the declarative description of a network's architecture.
-//! [`NetworkConfig`] bundles the per-sample input shape with an ordered stack of
-//! [`LayerConfig`]s, one per layer kind.
+//! [`LayerConfig`]: a typed, weight-free description of one layer.
 
 use crate::activations::Activation;
 use crate::layers::{Conv2d, Dense, Flatten, Layer};
@@ -9,82 +7,6 @@ use ndarray::{Ix2, Ix4};
 use ndarray_rand::rand::RngCore;
 use std::fmt;
 use std::sync::Arc;
-
-/// A network's architecture with no weights: the per-sample input shape and the ordered stack
-/// of layer configs — the declarative counterpart to a [`NeuralNetwork`](crate::model::NeuralNetwork).
-#[derive(Debug, Clone)]
-pub struct NetworkConfig {
-    /// The per-sample input shape, sample axis excluded.
-    pub input_shape: Vec<usize>,
-    /// The layers in order, from input to output.
-    pub layers: Vec<LayerConfig>,
-}
-
-impl NetworkConfig {
-    /// Starts a [`NetworkConfigBuilder`] at the given per-sample input shape.
-    pub fn builder(input_shape: Vec<usize>) -> NetworkConfigBuilder {
-        NetworkConfigBuilder::new(input_shape)
-    }
-}
-
-/// Fluent builder for a [`NetworkConfig`]: start from the per-sample input shape, append layers
-/// in order, then finish with [`build`](Self::build).
-pub struct NetworkConfigBuilder {
-    input_shape: Vec<usize>,
-    layers: Vec<LayerConfig>,
-}
-
-impl NetworkConfigBuilder {
-    /// Starts a builder at the given per-sample input shape, with no layers yet.
-    pub fn new(input_shape: Vec<usize>) -> Self {
-        Self {
-            input_shape,
-            layers: Vec::new(),
-        }
-    }
-
-    /// Appends a [`Conv2d`](LayerConfig::Conv2d) layer.
-    pub fn conv2d<A: Activation + 'static>(
-        mut self,
-        out_channels: usize,
-        kernel: (usize, usize),
-        stride: usize,
-        padding: usize,
-        activation: &Arc<A>,
-    ) -> Self {
-        self.layers.push(LayerConfig::Conv2d {
-            out_channels,
-            kernel,
-            stride,
-            padding,
-            activation: activation.clone(),
-        });
-        self
-    }
-
-    /// Appends a [`Flatten`](LayerConfig::Flatten) layer.
-    pub fn flatten(mut self) -> Self {
-        self.layers.push(LayerConfig::Flatten);
-        self
-    }
-
-    /// Appends a [`Dense`](LayerConfig::Dense) layer.
-    pub fn dense<A: Activation + 'static>(mut self, neurons: usize, activation: &Arc<A>) -> Self {
-        self.layers.push(LayerConfig::Dense {
-            neurons,
-            activation: activation.clone(),
-        });
-        self
-    }
-
-    /// Finishes the builder: the last added layer is the output.
-    pub fn build(self) -> NetworkConfig {
-        NetworkConfig {
-            input_shape: self.input_shape,
-            layers: self.layers,
-        }
-    }
-}
 
 /// A typed, weight-free description of one layer: its kind and the hyperparameters needed to
 /// build it, the concrete input shape excepted (that is threaded from the network's input).
@@ -313,7 +235,7 @@ impl std::error::Error for LayerConfigError {}
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::activations::{IDENTITY, RELU};
+    use crate::activations::RELU;
 
     use ndarray_rand::rand::SeedableRng;
     use ndarray_rand::rand::rngs::StdRng;
@@ -409,32 +331,6 @@ mod tests {
                 padding: 0,
             }
         );
-    }
-
-    #[test]
-    fn builder_assembles_layers_in_order() {
-        let config = NetworkConfig::builder(vec![1, 4, 4])
-            .conv2d(2, (3, 3), 1, 0, &RELU)
-            .flatten()
-            .dense(4, &RELU)
-            .dense(1, &IDENTITY)
-            .build();
-
-        assert_eq!(config.input_shape, vec![1, 4, 4]);
-        assert_eq!(config.layers.len(), 4);
-        assert!(matches!(config.layers[0], LayerConfig::Conv2d { .. }));
-        assert!(matches!(config.layers[1], LayerConfig::Flatten));
-        assert!(matches!(
-            config.layers[2],
-            LayerConfig::Dense { neurons: 4, .. }
-        ));
-        match &config.layers[3] {
-            LayerConfig::Dense {
-                neurons: 1,
-                activation,
-            } => assert_eq!(activation.name(), "identity"),
-            other => panic!("expected the head to be a Dense layer, got {other:?}"),
-        }
     }
 
     #[test]
